devtools/client/shared/components/reps/test/mochitest/test_reps_symbol.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Symbol rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - String</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
"use strict";
/* import-globals-from head.js */

window.onload = Task.async(function* () {
  const {
    REPS,
    getRep,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { SymbolRep } = REPS;

  let gripStubs = new Map();
  gripStubs.set("testSymbolFoo", {
    type: "symbol",
    name: "foo"
  });
  gripStubs.set("testSymbolWithoutIdentifier", {
    type: "symbol"
  });

  try {
    // Test that correct rep is chosen
    is(getRep(gripStubs.get("testSymbolFoo")), SymbolRep.rep,
      "Rep correctly selects SymbolRep Rep");

    // Test rendering
    yield testSymbol();
    yield testSymbolWithoutIdentifier();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testSymbol() {
    const renderedComponent = renderComponent(
      SymbolRep.rep,
      { object: gripStubs.get("testSymbolFoo") }
    );

    is(renderedComponent.textContent, "Symbol(foo)",
      "Symbol rep has expected text content");
  }

  function testSymbolWithoutIdentifier() {
    const renderedComponent = renderComponent(
      SymbolRep.rep,
      { object: gripStubs.get("testSymbolWithoutIdentifier") }
    );

    is(renderedComponent.textContent, "Symbol()",
      "Symbol rep without identifier has expected text content");
  }
});
</script>
</pre>
</body>
</html>
